requirements.txt
kagglehub
pandas
matplotlib
seaborn
numpy
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
transformers
gradio
